# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Cool API",
        "description": "API documentation for my Laravel application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sub-categories": {
            "get": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Get a paginated list of subcategories",
                "operationId": "getSubCategories",
                "parameters": [
                    {
                        "name": "paginate_count",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search subcategories by name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter subcategories by category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubCategory"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Create a new subcategory",
                "operationId": "storeSubCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Smartphones"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "smartphones"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mobile devices and accessories"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "SubCategory created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubCategory"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/sub-categories/{identifier}": {
            "get": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Get a single subcategory by ID or slug",
                "description": "Fetch a subcategory by its numeric ID or slug string.",
                "operationId": "showSubCategory",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "ID (integer) or slug (string) of the subcategory",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SubCategory fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Smartphones"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "smartphones"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Mobile phones and accessories"
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "category": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "SubCategory not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/sub-categories/{id}": {
            "post": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Update a subcategory",
                "description": "Updates an existing subcategory.",
                "operationId": "updateSubCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subcategory",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Method spoofing for PUT request.",
                                        "type": "string",
                                        "format": "hidden",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "updated-name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SubCategory updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SubCategory updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubCategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "SubCategory not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Delete a subcategory",
                "operationId": "deleteSubCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subcategory",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SubCategory deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SubCategory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Smartphones"
                    },
                    "slug": {
                        "type": "string",
                        "example": "smartphones"
                    },
                    "description": {
                        "type": "string",
                        "example": "Mobile devices and accessories"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-19T18:05:05.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-19T18:05:05.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}